move model classes from issuer-spi into dcp-spi
