+++ extensions/protocols/dcp/dcp-issuer/dcp-issuer-spi/build.gradle.kts
@@ -18,7 +18,6 @@ plugins {
 }
 
 dependencies {
-    api(project(":spi:verifiable-credential-spi"))
-    api(project(":spi:issuerservice:issuerservice-participant-spi"))
+    api(project(":extensions:protocols:dcp:dcp-spi"))
 }
