+++ extensions/protocols/dcp/dcp-spi/build.gradle.kts
@@ -18,6 +18,8 @@ plugins {
 }
 
 dependencies {
+    api(project(":spi:issuerservice:issuerservice-participant-spi"))
+    api(project(":spi:verifiable-credential-spi"))
     api(libs.jackson.annotation) // JsonProperty
 }
