+++ gradle/libs.versions.toml
@@ -42,6 +42,7 @@ edc-spi-vc = { module = "org.eclipse.edc:verifiable-credentials-spi", version.re
 edc-core-connector = { module = "org.eclipse.edc:connector-core", version.ref = "edc" }
 edc-core-token = { module = "org.eclipse.edc:token-core", version.ref = "edc" }
 edc-core-jersey = { module = "org.eclipse.edc:jersey-core", version.ref = "edc" }
+edc-core-api = { module = "org.eclipse.edc:api-core", version.ref = "edc" }
 edc-identity-did-core = { module = "org.eclipse.edc:identity-did-core", version.ref = "edc" }
 edc-identity-did-web = { module = "org.eclipse.edc:identity-did-web", version.ref = "edc" }
 edc-dcp-transform = { module = "org.eclipse.edc:identity-trust-transform", version.ref = "edc" }
